+++ contacts_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/bitcoin-sv/spv-wallet-go-client/fixtures"
-	"github.com/bitcoin-sv/spv-wallet/models"
+	"github.com/bitcoin-sv/spv-wallet/models/response"
 	responsemodels "github.com/bitcoin-sv/spv-wallet/models/response"
 	"github.com/stretchr/testify/require"
 )
